requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.28.57
botocore==1.31.57
colorama==0.4.6
coverage==7.2.2
Django==4.1.7
django-browser-reload==1.11.0
django-cleanup==8.0.0
django-cors-headers==3.14.0
django-filter==22.1
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.16
